services/data_processor/requirements.txt
bidict==0.22.0
certifi==2022.9.24
charset-normalizer==2.1.1
future==0.18.2
idna==3.4
iso8601==1.1.0
numpy==1.23.5
pyserial==3.5
python-dotenv==0.21.0
python-engineio==4.3.4
python-socketio==5.7.2
PyYAML==6.0
requests==2.28.1
urllib3==1.26.13
websocket-client==1.4.2
